+++ programs/spring/eveproxy/src/main/java/fr/guiguilechat/jcelechat/programs/spring/eveproxy/services/PlanetEvalService.java
@@ -18,8 +18,8 @@ import fr.guiguilechat.jcelechat.libs.spring.industry.blueprint.MaterialService;
 import fr.guiguilechat.jcelechat.libs.spring.industry.planetary.PlanetaryTaxService;
 import fr.guiguilechat.jcelechat.libs.spring.industry.planetary.SchemProductService;
 import fr.guiguilechat.jcelechat.libs.spring.industry.planetary.SchematicService;
-import fr.guiguilechat.jcelechat.libs.spring.items.type.Type;
-import fr.guiguilechat.jcelechat.libs.spring.items.type.TypeService;
+import fr.guiguilechat.jcelechat.libs.spring.sde.items.type.Type;
+import fr.guiguilechat.jcelechat.libs.spring.sde.items.type.TypeService;
 import fr.guiguilechat.jcelechat.libs.spring.trade.regional.MarketLine;
 import fr.guiguilechat.jcelechat.libs.spring.trade.regional.MarketLineService;
 import fr.guiguilechat.jcelechat.libs.spring.trade.tools.MaterialSourcing;
@@ -227,13 +227,15 @@ public class PlanetEvalService {
 			for (Entry<Integer, Long> e : fe.getProductById().entrySet()) {
 				long qtty = e.getValue();
 				Type t = typeService.byId(e.getKey());
-				volCost += params.getProductValuator().haulingCost(qtty, params.getVolumicPrice() * t.getVolume());
+				volCost += params.getProductValuator().haulingCost(qtty,
+						params.getVolumicPrice() * t.getVolume().doubleValue());
 				customTax += taxMult * exportTaxById.get(t.getId()) * qtty;
 			}
 			for (Entry<Integer, Long> e : fe.getMaterialsById().entrySet()) {
 				long qtty = e.getValue();
 				Type t = typeService.byId(e.getKey());
-				volCost += params.getMaterialSourcing().haulingCost(qtty, params.getVolumicPrice() * t.getVolume());
+				volCost += params.getMaterialSourcing().haulingCost(qtty,
+						params.getVolumicPrice() * t.getVolume().doubleValue());
 				customTax += taxMult * importTaxById.get(t.getId()) * qtty;
 			}
